requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-cors-headers==3.10.0
drf-yasg==1.20.0
gunicorn==20.1.0
arrow==1.2.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
model-bakery==1.3.3
pytest-factoryboy==2.1.0
pytest-mock==3.6.1
flake8==4.0.1
black==21.10b0